:truck: Rename ActivityFilter
